modification seed

+++ db/seeds.rb
@@ -6,10 +6,18 @@
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
 
-Item.destroy_all
-User.destroy_all
-Cart.destroy_all
-CartItem.destroy_all
+@i = Item.all
+CartItem.where(item_id: @i.ids).destroy_all
+@i.destroy_all
+@us = User.all
+@ca = Cart.all
+Order.where(cart_id:@ca.ids).destroy_all
+Cart.where(user_id:@us.ids).destroy_all
+@us.destroy_all
+
+
+
+
 
 picture_tab = ["https://cdn.pixabay.com/photo/2018/03/27/17/25/cat-3266673_960_720.jpg", "https://cdn.pixabay.com/photo/2018/10/01/09/21/pets-3715733_960_720.jpg", "https://cdn.pixabay.com/photo/2018/07/13/10/20/cat-3535404_960_720.jpg", "https://cdn.pixabay.com/photo/2017/03/14/14/49/cat-2143332_960_720.jpg", "https://cdn.pixabay.com/photo/2017/11/09/21/41/cat-2934720_960_720.jpg", "https://cdn.pixabay.com/photo/2014/11/30/14/11/kitty-551554_960_720.jpg", "https://cdn.pixabay.com/photo/2016/07/10/21/47/cat-1508613_960_720.jpg", "https://cdn.pixabay.com/photo/2016/03/28/12/35/cat-1285634_960_720.png", "https://cdn.pixabay.com/photo/2015/06/08/15/02/pug-801826_960_720.jpg", "https://cdn.pixabay.com/photo/2014/04/13/20/49/cat-323262_960_720.jpg", "https://cdn.pixabay.com/photo/2017/08/07/18/57/dog-2606759_960_720.jpg", "https://cdn.pixabay.com/photo/2015/02/25/17/56/cat-649164_960_720.jpg", "https://cdn.pixabay.com/photo/2019/06/18/11/32/cat-4282123_960_720.jpg", "https://cdn.pixabay.com/photo/2016/11/19/17/33/animal-1840495_960_720.jpg", "https://cdn.pixabay.com/photo/2016/11/29/13/52/animal-1870003_960_720.jpg", "https://cdn.pixabay.com/photo/2015/02/25/17/56/cat-649164_960_720.jpg"]
